Added new observer for lazy-loaded input fields
One some pages, the input fields may be lazy loaded due to React or other framework/practice. This means the event listener will not be added since that was done as page load. When the showFieldAttributes option is enabled ( ==true), we just rerun `updateShowFieldAttributesSettings()` which adds the event listeners for any of the newly added elements.